+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ## v1.73 Beta (04/09/2021)
 ### Added
 - Disable cards and automatic.
+- Added bar to show level progress.
+- Option to print cards and collectionables.
 
 ## v1.72 Beta (02/09/2021)
 ### Added